Move loading placeholder tests to shared location
