+++ src/integration/pages/onDemandTVPage/canonicalTests.js
@@ -1,9 +1,11 @@
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
+  runMediaPlaceholderTests,
 } from '../../common';
 
 export default () => {
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
+  runMediaPlaceholderTests();
 };
